clippy

+++ pallets/ajuna-awesome-avatars/src/types/avatar/versions/v3.rs
@@ -59,7 +59,7 @@ impl<T: Config> MinterV3<T> {
 			.map(|i| {
 				let (random_tier, random_variation) =
 					Self::random_component(season, hash, i as usize * 2, batched_mint);
-				((random_tier << 4) | random_variation) as u8
+				(random_tier << 4) | random_variation
 			})
 			.collect::<Vec<_>>();
 		Dna::try_from(dna).map_err(|_| Error::<T>::IncorrectDna.into())
